requirements.txt
langchain
streamlit
python-dotenv
PyPDF2
langchain-google-genai

-e .